feat: move fode systems to separate file
