+++ src/pycaputo/fode/gallery.py
@@ -1,0 +1,115 @@
+# SPDX-FileCopyrightText: 2024 Alexandru Fikl <alexfikl@gmail.com>
+# SPDX-License-Identifier: MIT
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+import numpy as np
+
+from pycaputo.logging import get_logger
+from pycaputo.typing import Array
+
+from .special import Function
+
+logger = get_logger(__name__)
+
+
+# {{{ Brusselator
+
+
+@dataclass(frozen=True)
+class Brusselator(Function):
+    r"""Implements the right-hand side of the Brusselator system.
+
+    .. math::
+
+        \begin{aligned}
+        D^\alpha[x](t) & =
+            a - (\mu + 1) x + x^2 y + A \cos \omega t, \\
+        D^\alpha[y](t) & =
+            \mu x - x^2 y.
+        \end{aligned}
+    """
+
+    a: float
+    """Parameter in the Brusselator model."""
+    mu: float
+    """Parameter in the Brusselator model."""
+
+    amplitude: float
+    """Forcing amplitude."""
+    omega: float
+    """Angular velocity of the forcing."""
+
+    if __debug__:
+
+        def __post_init__(self) -> None:
+            if self.a < 0:
+                raise ValueError(f"'a' must be positive: {self.a}")
+
+            if self.mu < 0:
+                raise ValueError(f"'mu' must be positive: {self.mu}")
+
+    def source(self, t: float, y: Array) -> Array:
+        f = self.amplitude * np.cos(self.omega * t)
+        return np.array([
+            self.a - (self.mu + 1) * y[0] + y[0] ** 2 * y[1] + f,
+            self.mu * y[0] - y[0] ** 2 * y[1],
+        ])
+
+    def source_jac(self, t: float, y: Array) -> Array:
+        return np.array([
+            [-(self.mu + 1) + 2 * y[0] * y[1], y[0] ** 2],
+            [self.mu - 2 * y[0] * y[1], -(y[0] ** 2)],
+        ])
+
+
+# }}}
+
+
+# {{{ van der Pol
+
+
+@dataclass(frozen=True)
+class VanDerPol(Function):
+    r"""Implements the right-hand side of the van der Pol system.
+
+    .. math::
+
+        \begin{aligned}
+        D^\alpha[x](t) & =
+            y, \\
+        D^\alpha[y](t) & =
+            \mu (1 - x^2) y - x + A \sin (\omega t).
+        \end{aligned}
+    """
+
+    mu: float
+    r"""Parameter indicating the strength of the nonlinearity and damping. Taking
+    :math:`\mu = -1` is common in the literature.
+    """
+
+    amplitude: float
+    """Forcing amplitude."""
+    omega: float
+    """Angular velocity of the forcing."""
+
+    if __debug__:
+
+        def __post_init__(self) -> None:
+            if self.mu < 0:
+                raise ValueError(f"'mu' must be positive: {self.mu}")
+
+    def source(self, t: float, y: Array) -> Array:
+        f = self.amplitude * np.cos(self.omega * t)
+        return np.array([y[1], self.mu * (1.0 - y[0] ** 2) * y[1] - y[0] + f])
+
+    def source_jac(self, t: float, y: Array) -> Array:
+        return np.array([
+            [0.0, 1.0],
+            [-2.0 * self.mu * y[0] * y[1] - 1.0, self.mu * (1.0 - y[0] ** 2)],
+        ])
+
+
+# }}}
